refactor: move books model server.js to /books/model.js working
