+++ src/books/model.js
@@ -1,0 +1,19 @@
+const { DataTypes } = require("sequelize");
+const connection = require("../db/connection");
+
+// Class on db entry info
+const Book = connection.define("Book", {
+    title: {
+        type: DataTypes.STRING,
+        unique: true,
+        allowNull: false,
+    },
+    author: {
+        type: DataTypes.STRING
+    },
+    genre: {
+        type: DataTypes.STRING
+    }
+});
+
+module.exports = Book;
